fix new points xp rating
