--- app/controllers/player_rating_controller.rb
@@ -1,2 +1,0 @@
-class PlayerRatingController < ApplicationController
-end
