--- app/helpers/player_rating_helper.rb
@@ -1,2 +1,0 @@
-module PlayerRatingHelper
-end
