+++ app/models/game_rating.rb
@@ -8,16 +8,10 @@ class GameRating < ActiveRecord::Base
   validates :activity, presence: true
 
   default_scope { order('date DESC, game_id DESC') }
+  scope :by_player, ->(player) { where("player_id = ?", player.id) }
+  scope :by_game_and_player, ->(game, player) { where("player_id = ? AND game_id = ?", player.id, game.id) }
 
   def value
     Rating.new(mean, deviation, activity)
-  end
-
-  def self.by_player(player)
-    self.where("player_id = ?", player.id)
-  end
-
-  def self.by_game_and_player(game, player)
-    self.where("player_id = ? AND game_id = ?", player.id, game.id)
   end
 end
