+++ app/models/xp_rating.rb
@@ -1,17 +1,10 @@
-include Saulabs::TrueSkill
-
-class MatchRating < ActiveRecord::Base
+class XpRating < ActiveRecord::Base
   belongs_to :player
   belongs_to :match
-  validates :mean, presence: true
-  validates :deviation, presence: true
-  validates :activity, presence: true
-
+  validates :value, presence: true
+  
   default_scope { order('date DESC, match_id DESC') }
   scope :by_player, ->(player) { where("player_id = ?", player.id) }
   scope :by_match_and_player, ->(match, player) { where("player_id = ? AND match_id = ?", player.id, match.id) }
-
-  def value
-    Rating.new(mean, deviation, activity)
-  end
+  scope :by_player_and_date, ->(player, date) { where("player_id = ? AND date >= ? AND date <= ?", player.id, date.beginning_of_day, date.end_of_day)}
 end
