+++ db/migrate/20150329182640_create_match_ratings.rb
@@ -3,7 +3,7 @@ class CreateMatchRatings < ActiveRecord::Migration
     create_table :match_ratings do |t|
       t.integer :player_id, null: false
       t.integer :match_id, null: false, default: 0
-      t.float :mean, null: false, default: 25.0
+      t.float :mean, null: false, default: 50.0
       t.float :deviation, null: false, default: 2.0
       t.float :activity, null: false, default: 1.0
