+++ db/migrate/20150329182646_create_game_ratings.rb
@@ -3,7 +3,7 @@ class CreateGameRatings < ActiveRecord::Migration
     create_table :game_ratings do |t|
       t.integer :player_id, null: false
       t.integer :game_id, null: false, default: 0
-      t.float :mean, null: false, default: 25.0
+      t.float :mean, null: false, default: 50.0
       t.float :deviation, null: false, default: 2.0
       t.float :activity, null: false, default: 1.0
