+++ db/migrate/20150330001037_drop_player_ratings.rb
@@ -1,6 +1,5 @@
 class DropPlayerRatings < ActiveRecord::Migration
   def change
     drop_table :player_ratings
-    Player.all.each { |p| p.update_attribute(:trueskill, p.calculate_trueskill) }
   end
 end
